Refactored sorters and formatters, deleted impure prop setters from DataTable, and deleted redundant custom formatters for date objects
