+++ src/components/ComponentDemos.js
@@ -5,9 +5,9 @@ import styled from 'styled-components/macro'
 
 import { getTypes } from '../utils/api'
 import { ReportModal } from './form/ReportModal'
+import ImportsTable from './table/ImportsTable'
 import Button from './ui/Button'
 import CaptionInput from './ui/CaptionInput'
-import ImportsTable from './ui/ImportsTable'
 import Input from './ui/Input'
 import ListEntry from './ui/ListEntry'
 import ProgressBar from './ui/ProgressBar'
